Add hover effect
Wanted to test out if I could make a hover effect when hovering
over an image. However I realized that what I actually want is for
all the other images to be blurred out so that the person can focus
on the non-blurred image. I just want to commit now in case I make
a mistake.